requirements.txt
asgiref==3.6.0
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.2
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.32.0
colorama==0.4.6
constantly==15.1.0
creategitignore==0.2.1
cryptography==41.0.1
daphne==4.0.0
dj-database-url==2.0.0
Django==4.1.7
django-autoslug==1.9.8
django-cloudinary-storage==0.3.0
django-compat==1.0.15
django-cors-headers==3.14.0
django-countries==7.5.1
django-filter==23.1
django-jazzmin==2.6.0
django-rest-framework==0.1.0
django-shortcuts==1.6
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
exceptiongroup==1.1.1
Faker==17.6.0
filestorage==0.0.6
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
gunicorn==20.1.0
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.1.0
pluggy==1.0.0
protobuf==4.22.1
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
python-flutterwave==0.8.10
pytz==2022.7.1
redis==4.5.5
requests==2.28.2
rsa==4.9
service-identity==23.1.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.4
tomli==2.0.1
treblle==0.14
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
utils==1.0.1
whitenoise==6.4.0
zope.interface==6.0
